+++ src/main/java/com/lali/financial/habits/management/service/repository/IncomeRepository.java
@@ -13,5 +13,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IncomeRepository extends JpaRepository<Income, Integer> {
+public interface IncomeRepository extends JpaRepository<Income, Long> {
 }
